calculator/pyproject.toml
[project]
name = "calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "matplotlib>=3.10.8",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
